Lab2

+++ Lab/Lab1/Lab1_1913769/Programming/Problem3/factorial.c
@@ -1,6 +1,6 @@
 #include "factorial.h"
 
-int factorial(const int aNumber) {
+long long int factorial(const long int aNumber) {
 	if(aNumber == 1) return 1;
 	return aNumber * factorial(aNumber-1);
 }
