+++ Lab/Lab1/Lab1_1913769/Programming/Problem3/factorial.h
@@ -3,6 +3,6 @@
 #ifndef FACTORIAL_H
 #define FACTORIAL_H
 
-int factorial(const int aNumber);
+long long int factorial(const long int aNumber);
 
 #endif
